[CLN/WIP] Try to add preconditioners
